feat: on peut faire bouger la boule avec z,q,s,d

+++ src/volume.c
@@ -130,7 +130,7 @@ void dessiner_maison()
     
 }
 
-void dessiner_boule(int rayon, int x, int y, int z)
+void dessiner_boule(float rayon, float x, float y, float z)
 {
     glColor3f(1,0,0);   //Rouge 
     glTranslatef(x, y, z);
@@ -138,3 +138,21 @@ void dessiner_boule(int rayon, int x, int y, int z)
     glTranslatef(-x, -y, -z);
     glColor3f(1,1,1);   //Blanc
 }
+
+//   glClearColor(0.0F,0.0F,0.0F,0.0F) ;
+//   glClear(GL_COLOR_BUFFER_BIT|GL_DEPTH_BUFFER_BIT) ;
+//   glPushMatrix() ;
+//   glRotatef(anglex,1.0F,0.0F,0.0F) ;
+//   glRotatef(angley,0.0F,1.0F,0.0F) ;
+//   glRotatef(anglez,0.0F,0.0F,1.0F) ;
+//   glTranslatef(0.5F,0.0F,0.0F) ;
+//   GLfloat rouge[] = { 1.0,0.0,0.0,0.0 };
+//   glMaterialfv(GL_FRONT,GL_DIFFUSE,rouge);
+//   glutSolidCube(0.6) ;
+//   glTranslatef(-1.0F,0.0F,0.0F) ;
+//   GLfloat vert[] = { 0.0,1.0,0.0,0.0 };
+//   glMaterialfv(GL_FRONT,GL_DIFFUSE,vert);
+//   glutSolidSphere(0.4,50,50) ;
+//   glPopMatrix() ;
+//   glFlush() ;
+//   glutSwapBuffers() ;
